Me when I fix literally every assignment

+++ Lecture10/a10_p6/Critter.h
@@ -33,7 +33,7 @@ class Critter
     void setThirst(double newthirst);
     // getter method
     int getHunger();
-    int getThirst();
+    double getThirst();
     // service method
     void print();
 };
